Fix default allowed url
