+++ src/main/java/io/hexlet/typoreporter/handler/exception/ForbiddenDomainException.java
@@ -11,6 +11,10 @@ public class ForbiddenDomainException extends ErrorResponseException {
     private static final String MESSAGE = "Domain ''{0}'' is forbidden";
 
     public ForbiddenDomainException(final String domain) {
-        super(FORBIDDEN, ProblemDetail.forStatusAndDetail(FORBIDDEN, "Domain is forbidden"), null, format(MESSAGE, domain), new Object[]{domain});
+        super(FORBIDDEN, ProblemDetail.forStatusAndDetail(FORBIDDEN,
+                                        "Domain is forbidden"),
+                                        null,
+                                        format(MESSAGE, domain),
+                                        new Object[]{domain});
     }
 }
