+++ src/main/resources/messages_en.properties
@@ -93,9 +93,8 @@ text.created-at=Created at
 text.modified-at=Modified at
 text.reported-by=Reported by
 text.add-user-to-wks=Adding a user to a workspace
-text.add-allowed-url-to-wks=Add another URL from which reports can be sent to the workspace
-text.main-url=Main URL for reports is
-text.input-url-placeholder=Enter URL. For example: https://fixit.hexlet.io/
+text.allowed-urls=URLs from which reports can be sent to the workspace:
+text.input-url-placeholder=Add another allowed URL. For example: https://fixit.hexlet.io/
 text.confirm-delete-user-wks=Are you sure you want to delete user from this workspace?
 text.confirm-delete-url-wks=Are you sure you want to delete allowed url from this workspace?
 text.hint-choose-a-user=Choose a user from the list
