+++ src/main/resources/messages_ru.properties
@@ -91,9 +91,9 @@ text.created-at=Создан
 text.modified-at=Изменен
 text.reported-by=Отправитель
 text.add-user-to-wks=Добавить пользователя в пространство
-text.add-allowed-url-to-wks=Добавить дополнительный домен, с которого можно отправлять отчёты в пространство
-text.main-url=Основной URL для отчётов -
-text.input-url-placeholder=Введите URL. Например: https://fixit.hexlet.io/
+text.allowed-urls=Домены, с которых можно отправлять отчёты в пространство:
+text.add-allowed-url-to-wks=Добавить дополнительный домен
+text.input-url-placeholder=Введите дополнительный домен. Например: https://fixit.hexlet.io/
 text.confirm-delete-user-wks=Вы действительно хотите удалить пользователя из рабочего пространства?
 text.confirm-delete-url-wks=Вы действительно хотите удалить домен для отчётов из рабочего пространства?
 text.hint-choose-a-user=Выберите пользователя из списка
