--- src/main/resources/templates/workspace/wks-allowed-urls.html
@@ -1,130 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:sd="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
-<head th:replace="~{fragments/header :: head}"></head>
-<body style="padding-top: 4.5rem">
-<nav th:replace="~{fragments/panels :: mainNavbarTop}"></nav>
-<main class="container-fluid">
-    <div class="row">
-        <!-- Workspace Side Bar -->
-        <div class="col-2" th:include="fragments/panels :: workspaceSideBar (${wksInfo.id()}, ${wksName}, 'urls')"></div>
-        <div class="col">
-            <!-- Workspace progress typo bar -->
-            <div class="row my-3"
-                 th:if="${sumTypoInWks} > 0"
-                 th:insert="~{fragments/workspace :: wksProgressBar (${sumTypoInWks}, ${countTypoByStatus})}"></div>
-            <!-- Workspace number of typo -->
-            <div class="row">
-                <p th:text="|#{text.number-of-typos(${wksName}, ${sumTypoInWks})}|"></p>
-            </div>
-            <!-- Main url of the workspace -->
-            <div class="row">
-                <div class="col">
-                    <h4>
-                        [[#{text.main-url}]] <small class="text-muted" th:text="${mainUrl}"></small>
-                        <a th:href="@{'/workspace/' + ${wksId} + '/update'}" class="btn btn-light" th:text="#{btn.edit}"></a>
-                    </h4>
-                </div>
-            </div>
-            <!-- Adding an url to a workspace -->
-            <div class="row my-3">
-                <p th:text="#{text.add-allowed-url-to-wks}"></p>
-                <div class="col">
-                    <form method="POST" th:action="@{'/workspace/' + ${wksId} + '/allowed-urls'}" th:object="${inputUrl}">
-                        <div class="row">
-                            <div class="col-7">
-                                <input type="url" class="form-control mb-2 col" id="inputUrl" th:placeholder="#{text.input-url-placeholder}" th:field="*{url}"
-                                       th:classappend="${#fields.hasErrors('url') && formModified}? 'is-invalid'"
-                                       th:errorclass="is-invalid" required>
-                            </div>
-                            <div class="col-5">
-                                <button type="submit" class="btn btn-light" th:text="#{btn.add-to-wks}"></button>
-                            </div>
-                        </div>
-                        <div class="row mt-2">
-                            <div class="col-12">
-                                <ul class="alert alert-danger ps-5" th:if="${#fields.hasErrors('url')}">
-                                    <li th:each="err : ${#fields.errors('url')}" th:text="${err}"></li>
-                                </ul>
-                            </div>
-                        </div>
-                    </form>
-                </div>
-            </div>
-            <!-- Workspace typo page -->
-            <div class="row" th:if="${urlsPage.hasContent()}">
-                <main class="col">
-                    <div class="row">
-                        <!-- Page size dropdown -->
-                        <div class="col-5">
-                            <div class="dropdown">
-                                <button aria-expanded="false"
-                                        class="btn btn-light dropdown-toggle"
-                                        data-bs-toggle="dropdown"
-                                        id="dropdownSizePage"
-                                        th:text="|#{btn.page-size} ${urlsPage.size}|"
-                                        type="button">
-                                </button>
-                                <ul aria-labelledby="dropdownSizePage" class="dropdown-menu">
-                                    <li th:each="availableSize : ${availableSizes}">
-                                        <a class="dropdown-item"
-                                           th:classappend="(${urlsPage.size} == ${availableSize} ? 'disabled' : '')"
-                                           th:href="@{'/workspace/' + ${wksId} + '/allowed-urls'(page=0,size=${availableSize},${sortDir}|)}"
-                                           th:text="${availableSize}">
-                                        </a>
-                                    </li>
-                                </ul>
-                            </div>
-                        </div>
-                        <!-- Page pagination -->
-                        <div class="col-sm-6">
-                            <ul class="pagination" sd:pagination="full"></ul>
-                        </div>
-                    </div>
-                    <!-- Workspace urls table -->
-                    <div class="row">
-                        <div class="col">
-                            <table class="table">
-                                <!-- Workspace urls table head -->
-                                <colgroup>
-                                    <col class="col-11">
-                                    <col class="col-1">
-                                </colgroup>
-                                <thead th:object="${urlsPage}">
-                                    <tr>
-                                        <th scope="col" colspan="2">
-                                            <a class="btn btn-light"
-                                               th:href="@{'/workspace/' + ${wksId} + '/allowed-urls'(size=*{size},sort=|url,${sortDir.isAscending() ? DESC : ASC}|)}">
-                                                <span th:text="#{report-url}"></span> <span>[[${sortDir.isAscending() ? '&uArr;' : '&dArr;'}]]</span>
-                                            </a>
-                                        </th>
-                                    </tr>
-                                </thead>
-                                <tbody>
-                                    <th:block th:each="url : ${urlsPage.content}" th:object="${url}">
-                                        <!-- Workspace url table row -->
-                                        <tr>
-                                            <td class="align-middle" th:text="*{url}"></td>
-                                            <td>
-                                                <form th:method="DELETE"
-                                                      th:if="isAdmin"
-                                                      th:action="@{'/workspace/' + ${wksId} + '/allowed-urls'}">
-                                                    <input type="hidden" th:value="*{url}" name="url">
-                                                    <button type="submit" class="btn btn-danger"
-                                                            th:text="#{btn.delete}"
-                                                            th:onclick="return confirm(/*[[#{text.confirm-delete-url-wks}]]*/)"
-                                                    ></button>
-                                                </form>
-                                            </td>
-                                        </tr>
-                                    </th:block>
-                                </tbody>
-                            </table>
-                        </div>
-                    </div>
-                </main>
-            </div>
-        </div>
-    </div>
-</main>
-</body>
-</html>
